refactor(worker): rename module with consistent naming scheme
